Help: move the "Vienna Help" directory to "Vinkix Help". I should have done this before editing the help files.
The real problem was that I missed a string "Vienna Help" in the index.html file.  Correcting that means
that the help index screen now display correctly. It is still in the "Vienna Help" direcotry. 

See check-in [9eb3f55b] . This commit is now a closed branch.  Continue working on the 
rename-to-vinkix2 branch.